requirements.txt
adbc_driver_manager==1.11.0
adbc_driver_postgresql==1.11.0
adbc_driver_sqlite==1.11.0
AppKit==0.2.8
argcomplete==3.6.3
array_api_strict==2.5
astroid==4.1.2
asyncssh_unofficial==0.9.2
BeaEnginePython==3.1.0
beautifulsoup4==4.14.3
black==26.3.1
bokeh==3.9.0
botocore==1.42.90
brotli==1.2.0
brotlicffi==1.2.0.1
cairocffi==1.7.1
capstone==5.0.7
Carbon==1.1.10
cerealizer==0.8.4
clr==1.0.3
colorama==0.4.6
ConfigParser==7.2.0
coverage==7.13.5
curio==1.6
Cython==3.2.4
defcon==0.12.2
defusedxml==0.7.1
distorm3==3.5.2
dl==0.1.0
docrepr==0.2.0
fastparquet==2026.3.0
Foundation==0.1.0a0.dev1
freetype_py==2.5.1
fs==2.4.16
fsspec==2026.3.0
gevent==26.4.0
glyphsLib==6.13.0
gobject==0.1.0
greenlet==3.4.0
h2==4.3.0
HTMLParser==0.0.2
hypothesis==6.152.1
importlib_resources==7.1.0
inflect==7.5.0
ini2toml==0.15
ipyparallel==9.1.0
ipywidgets==8.1.8
jax==0.10.0
jnius==1.1.0
jupyter_ai==3.0.0
jupyter_ai_magics==2.31.7
keyring==25.7.0
lxml==6.0.4
lz4==4.4.5
mpmath==1.4.1
mtrand==0.1
munkres==1.1.4
mypy==1.20.1
nbformat==5.10.4
ndonnx==0.18.0
netifaces==0.11.0
nose==1.3.7
numexpr==2.14.1
numpydoc==1.10.0
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
orjson==3.11.8
pooch==1.9.0
psyco==1.6
py==1.11.0
pycares==5.0.1
pycurl==7.45.7
pyczmq==0.0.4
PyInstaller==6.19.0
pyodide==0.0.2
PyOpenGL==3.1.10
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.11.0
PySide==1.2.4
Pyside==1.2.4
PySide2==5.15.2.1
PySide6==6.11.0
pytest_fail_slow==0.6.0
pytest_run_parallel==0.8.2
pytest_timeout==2.4.0
python_calamine==0.6.2
pyxlsb==1.0.10
qtpy==2.4.3
Quartz==0.0.1.dev0
railroad==0.5.0
regex==2026.4.4
reportlab==4.4.10
Res==0.1.7
s3fs==2026.3.0
scikit_learn==1.8.0
scikits.talkbox==0.2.5
scipy_doctest==2.2.0
selenium==4.43.0
setuptools_scm==10.0.5
shiboken2==5.15.2.1
shiboken6==6.11.0
sip==6.15.3
Sphinx==9.1.0
sphinx_gallery==0.20.0
SQLAlchemy==2.0.49
stackless==0.0.1
sympy==1.14.0
testpath==0.6.0
threadpoolctl==3.6.0
trio==0.33.0
trove_classifiers==2026.1.14.14
typing_extensions==4.15.0
ufoLib2==0.18.1
uharfbuzz==0.53.6
unicodedata2==17.0.1
urllib3_secure_extra==0.1.0
xattr==1.3.0
xlrd==2.0.2
xlsxwriter==3.2.9
xmlrpclib==1.0.1
xx==3.3.2
yapf==0.43.0
zopfli==0.4.1
